Bug fixes and stuff

+++ KitchenApp/KitchenLib/OwnedProduct.cs
@@ -5,14 +5,14 @@ namespace KitchenLib
     public class OwnedProduct : Product
     { 
         public DateTime _consume_before { get; set; }
-        public uint _stock { get; set; }
+        public long _stock { get; set; }
         public string _owner_uid { get; set; }
 
         public OwnedProduct()
         {
         }
 
-        public uint restock(uint quantity)
+        public long restock(long quantity)
         {
             _quantity += quantity;
             return _quantity;
